1and2 complete

+++ src/Main.java
@@ -4,21 +4,20 @@ import java.nio.file.Paths;
 import java.util.*;
 import java.util.zip.*;
 public class Main {
-
     static StringBuilder sb = new StringBuilder();
     public static void main(String[] args) {
 
-        // For first task
+        // Первая задача
         List<String> nameSubDir = Arrays.asList("src", "res", "savegames", "temp");
         List<String> nameSupDir1 = Arrays.asList("main", "test");
         List<String> nameSupDir2 = Arrays.asList("drawables", "vectors", "icons");
         List<String> nameFile1 = Arrays.asList("Main.java", "Utils.java");
         List<String> nameFile2 = Arrays.asList("temp.txt");
 
-        makeDir(nameSubDir, nameSupDir1, nameSupDir2);
-        makeFile(nameFile1, nameFile2);
+        makeDir(nameSubDir, nameSupDir1, nameSupDir2); // Создаем каталоги
+        makeFile(nameFile1, nameFile2); // Создаем файлы
 
-        // For second task
+        // Вторая задача
         GameProgress gp1 = new GameProgress(100,100,100,100);
         GameProgress gp2 = new GameProgress(55,55,55,55);
         GameProgress gp3 = new GameProgress(4,4,4,4);
@@ -26,25 +25,19 @@ public class Main {
         String link2 = "C://Workspace_Java//Temp//Games//savegames//save2.dat";
         String link3 = "C://Workspace_Java//Temp//Games//savegames//save3.dat";
 
-        saveGame(link1, gp1);
-        saveGame(link2, gp2);
-        saveGame(link3, gp3);
+        saveGame(link1, gp1); // Сохраняем первую игру
+        saveGame(link2, gp2); // Сохраняем вторую игру
+        saveGame(link3, gp3); // Сохраняем третью игру
 
         String linkZip = "C://Workspace_Java//Temp//Games//savegames//zip.zip";
         List<String> list = Arrays.asList(link1, link2, link3);
 
-        zipFiles(linkZip, list);
+        zipFiles(linkZip, list); // Добавляем в архив сохранения
 
-        try (FileWriter writer = new FileWriter("C://Workspace_Java//Temp//Games//temp//temp.txt", true))
-        {
-            writer.write(sb.toString());
-        }
-        catch (IOException ex) {
-            System.out.println(ex.getMessage());
-        }
+        delFile(); // Удаляем файлы
 
+        log(); // Записываем логи
     }
-
     // Создаем каталоги
     static void makeDir(List<String> list1, List<String> list2, List<String> list3) {
 
@@ -55,7 +48,6 @@ public class Main {
             } else {
                 sb.append("Direct ").append(s).append(" don't create\n");
             }
-
         }
         for (String s : list2) {
             File dir = new File("C://Workspace_Java//Temp//Games//src//" + s);
@@ -74,7 +66,6 @@ public class Main {
             }
         }
     }
-
     // Создаем файлы
     static void makeFile(List<String> list1, List<String> list2) {
 
@@ -99,49 +90,58 @@ public class Main {
         }
         catch (IOException ex) {System.out.println(ex.getMessage());}
     }
-
-    // Создаем файлы сохранений
+    // Записываем сохранения
     static void saveGame(String link, GameProgress gp) {
+
         File file = new File(link);
-        try(ObjectOutputStream fos = new ObjectOutputStream(Files.newOutputStream(Paths.get(link))))
+        try(ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(link)))
         {
             file.createNewFile();
             byte[] buffer = gp.toString().getBytes();
-            fos.write(buffer, 0, buffer.length);
+            oos.writeObject(buffer);
+            oos.close();
         }
-        catch(Exception ex){
-
-            System.out.println(ex.getMessage());
-        }
+        catch(IOException ex){System.out.println(ex.getMessage());}
     }
-
     // Архивируем файлы сохранения и удаляем исходники
     static void zipFiles(String linkZip, List<String> list) {
 
-        for(String link : list) {
-            try (ZipOutputStream zout = new ZipOutputStream(Files.newOutputStream(Paths.get(linkZip)));
-                FileInputStream fis = new FileInputStream(link);)
+            try (ZipOutputStream zout = new ZipOutputStream(new FileOutputStream(linkZip)))
             {
-                File file = new File(link);
-                ZipEntry entry1 = new ZipEntry(file.getName());
-                zout.putNextEntry(entry1);
-
-                byte[] buffer = new byte[fis.available()];
-                fis.read(buffer);
-
-                zout.write(buffer);
-                sb.append("Zip ").append(file.getName()).append(" create\n");
-                zout.closeEntry();
-                fis.close();
-
-                boolean deleted = file.delete();
-                if(deleted)
-                    sb.append("File ").append(file.getName()).append(" has been deleted\n");
+                for(String link : list) {
+                    FileInputStream fis = new FileInputStream(link);
+                    File file = new File(link);
+                    zout.putNextEntry(new ZipEntry(file.getName()));
+                    byte[] buffer = new byte[fis.available()];
+                    fis.read(buffer, 0, buffer.length);
+                    zout.write(buffer);
+                    sb.append("Zip ").append(file.getName()).append(" create\n");
+                    zout.closeEntry();
+                    fis.close();
+                }
 
             }
-            catch (Exception ex) {
-                System.out.println(ex.getMessage());
+            catch (Exception ex) {System.out.println(ex.getMessage());}
+    }
+    // Удаляем файлы
+    static void delFile() {
+        File dir = new File("C://Workspace_Java//Temp//Games//savegames");
+        List<File> files= Arrays.asList(dir.listFiles());
+        for (File file : files) {
+            if(!file.getName().equals("zip.zip")) {
+                file.delete();
+                sb.append("File ").append(file.getName()).append(" has been deleted\n");
             }
+        }
+    }
+    // Запись логов
+    static void log() {
+        try (FileWriter writer = new FileWriter("C://Workspace_Java//Temp//Games//temp//temp.txt", true))
+        {
+            writer.write(sb.toString());
+        }
+        catch (IOException ex) {
+            System.out.println(ex.getMessage());
         }
     }
 }
